Use full weekday name in Clock09 footer

+++ clock-variants.js
@@ -50,7 +50,7 @@
   function nowParts() {
     const p = new Intl.DateTimeFormat('en-AU', {
       timeZone: TZ, hour12: false,
-      weekday:'short', day:'2-digit', month:'short', year:'numeric',
+      weekday:'long', day:'2-digit', month:'short', year:'numeric',
       hour:'2-digit', minute:'2-digit', second:'2-digit'
     }).formatToParts(new Date());
     const out = {};
